pyproject.toml
[project]
name = "clawbridge"
version = "0.1.0"
description = "Build OpenClaw-style agents inside frameworks like Agno and Agentica"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
agno = [
    "agno>=1.1",
    "fastapi",
    "uvicorn",
    "anthropic",
    "openai",
    "groq",
    "sqlalchemy>=2.0",
    "psycopg>=3.0",
    "pgvector>=0.3",
    "lancedb>=0.24",
    "pyarrow>=18.0",
    "slack-sdk>=3.0",
]
agentica = ["symbolica-agentica>=0.3"]
all = ["clawbridge[agno,agentica]"]
dev = ["pytest>=8.0", "ruff>=0.8", "mypy>=1.13", "types-PyYAML"]

[project.scripts]
clawbridge = "clawbridge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
